+++ django_forest/utils/views/action.py
@@ -1,17 +1,11 @@
-import json
-
 from django.http import HttpResponse
-from django.views import generic
 
 from django_forest.utils import get_token
 from django_forest.utils.permissions import Permission
+from .base import BaseView
 
 
-class ActionView(generic.View):
-    def get_body(self, body):
-        body_unicode = body.decode('utf-8')
-        return json.loads(body_unicode)
-
+class ActionView(BaseView):
     def get_smart_action_request_info(self, request):
         return {
             'endpoint': request.path,
@@ -30,13 +24,19 @@ class ActionView(generic.View):
 
     def dispatch(self, request, *args, **kwargs):
         body = self.get_body(request.body)
-        collection_name = body['data']['attributes']['collection_name']
-        # check permissions
+        resource = body['data']['attributes']['collection_name']
+
         try:
-            token = get_token(request)
-            smart_action_request_info = self.get_smart_action_request_info(request)
-            self.is_authorized(collection_name, token, smart_action_request_info)
-        except Exception:
-            return HttpResponse(status=403)
+            self.Model = self.get_model(resource)
+        except Exception as e:
+            return self.error_response(e)
         else:
-            return super().dispatch(request, *args, **kwargs)
+            # check permissions
+            try:
+                token = get_token(request)
+                smart_action_request_info = self.get_smart_action_request_info(request)
+                self.is_authorized(resource, token, smart_action_request_info)
+            except Exception:
+                return HttpResponse(status=403)
+            else:
+                return super().dispatch(request, *args, **kwargs)
